22-07-22-0235

+++ src/users/users.service.ts
@@ -1,10 +1,17 @@
-import { AuthRegisterDto } from '../auth/dto/auth-register.dto';
-import { PrismaService } from './../prisma/prisma.service';
-import { Injectable, ForbiddenException } from '@nestjs/common';
-import { UpdateUserDto } from './dto/update-user.dto';
-import { prismaExclude } from './../helper/prisma_exclude';
+// checked
+import {
+  Injectable,
+  ForbiddenException,
+  HttpStatus,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaClientKnownRequestError } from '@prisma/client/runtime';
 import { AuthService } from 'src/auth/auth.service';
+import { PrismaService } from './../prisma/prisma.service';
+import { AuthRegisterDto } from '../auth/dto/auth-register.dto';
+import { UpdateUserDto } from './dto/update-user.dto';
+import { prismaExclude } from './../helper/prisma_exclude';
+
 import * as argon from 'argon2';
 
 @Injectable()
@@ -14,29 +21,31 @@ export class UsersService {
     private authService: AuthService,
   ) {}
 
-  // POST /users
-  async create(authSignUpDto: AuthRegisterDto) {
+  async create(dto: AuthRegisterDto) {
     try {
-      return await this.authService.register(authSignUpDto);
+      return await this.authService.register(dto);
     } catch (error) {
       throw error;
     }
   }
 
-  // GET /users
   async findAll() {
     try {
       const user = await this.prisma.user.findMany({
         select: prismaExclude.$exclude('user', ['hashed_password']),
       });
 
-      return user;
+      return {
+        status: HttpStatus.OK,
+        success: true,
+        message: 'Berhasil mengambil semua data user',
+        data: user,
+      };
     } catch (error) {
       throw error;
     }
   }
 
-  // GET /users/id
   async findOne(id: number) {
     try {
       const user = await this.prisma.user.findUnique({
@@ -45,18 +54,27 @@ export class UsersService {
       });
 
       if (!user) {
-        throw new ForbiddenException(`User dengan id : ${id} tidak ditemukan`);
+        throw new NotFoundException({
+          status: HttpStatus.NOT_FOUND,
+          success: false,
+          message: `Gagal mengambil data, user dengan id : ${id} tidak ditemukan`,
+          data: {},
+        });
       }
 
-      return user;
+      return {
+        status: HttpStatus.FOUND,
+        success: true,
+        message: `Berhasil mengambil data user dengan id : ${id}`,
+        data: user,
+      };
     } catch (error) {
       throw error;
     }
   }
 
-  // PATCH /users/id
-  async update(id: number, updateUserDto: UpdateUserDto) {
-    const { name, username, password, is_active } = updateUserDto;
+  async update(id: number, dto: UpdateUserDto) {
+    const { name, username, password } = dto;
     const hashed_password = await argon.hash(password);
 
     try {
@@ -66,22 +84,34 @@ export class UsersService {
           name,
           username,
           hashed_password,
-          is_active,
         },
         select: prismaExclude.$exclude('user', ['hashed_password']),
       });
 
-      return user;
+      return {
+        status: HttpStatus.OK,
+        success: true,
+        message: `Berhasil mengubah data user dengan id : ${id}`,
+        data: user,
+      };
     } catch (error) {
       if (error instanceof PrismaClientKnownRequestError) {
         if (error.code === 'P2025') {
-          throw new ForbiddenException(
-            `User dengan id : ${id} tidak ditemukan`,
-          );
+          throw new NotFoundException({
+            status: HttpStatus.NOT_FOUND,
+            success: false,
+            message: `Gagal mengubah, user dengan id : ${id} tidak ditemukan`,
+            data: {},
+          });
         }
 
         if (error.code === 'P2002') {
-          throw new ForbiddenException(`Username ${username} sudah terpakai`);
+          throw new ForbiddenException({
+            status: HttpStatus.FORBIDDEN,
+            success: false,
+            message: `Username ${username} sudah terpakai`,
+            data: {},
+          });
         }
       }
 
@@ -89,27 +119,42 @@ export class UsersService {
     }
   }
 
-  // DELETE /users/id
   async remove(id: number) {
     try {
+      const role = await this.findOne(id);
+      if (role.data.role === 'admin') {
+        throw new ForbiddenException({
+          status: HttpStatus.FORBIDDEN,
+          success: false,
+          message: `Tidak diizinkan untuk menghapus user dengan role Admin`,
+          data: {},
+        });
+      }
+
       const user = await this.prisma.user.delete({
         where: { id },
         select: prismaExclude.$exclude('user', ['hashed_password']),
       });
 
-      return user;
+      return {
+        status: HttpStatus.OK,
+        success: true,
+        message: `Berhasil menghapus user dengan id : ${id}`,
+        data: user,
+      };
     } catch (error) {
       if (error instanceof PrismaClientKnownRequestError) {
         if (error.code === 'P2025') {
-          throw new ForbiddenException(
-            `User dengan id : ${id} tidak ditemukan`,
-          );
+          throw new NotFoundException({
+            status: HttpStatus.NOT_FOUND,
+            success: false,
+            message: `Gagal menghapus, user dengan id : ${id} tidak ditemukan`,
+            data: {},
+          });
         }
-
-        return error;
       }
 
-      // throw error;
+      throw error;
     }
   }
 }
